fixes for various reducers
