+++ src/reducers/total_transactions_count_by_contract_addresses.rs
@@ -19,7 +19,7 @@ impl Reducer {
     ) -> Result<(), gasket::error::Error> {
         let key = match &self.config.key_prefix {
             Some(prefix) => prefix.to_string(),
-            None => "total_transactions_count_by_addresses".to_string(),
+            None => "total_transactions_count_by_contract_addresses".to_string(),
         };
 
         let crdt = model::CRDTCommand::PNCounter(key, 1);
@@ -50,6 +50,6 @@ impl Reducer {
 impl Config {
     pub fn plugin(self) -> super::Reducer {
         let reducer = Reducer { config: self };
-        super::Reducer::TotalTransactionsCountByAddresses(reducer)
+        super::Reducer::TotalTransactionsCountByContractAddresses(reducer)
     }
 }
